requirements.txt
asgiref==3.8.1
Django==5.0.12
django-extensions==3.2.3
djangorestframework==3.15.2
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.1
